+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## UNRELEASED
 
+IMPROVEMENTS:
+ * agent: Update `hashicorp/nomad/api` to v1.3.1 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * agent: Update `armon/go-metrics` to v0.3.11 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * plugin/apm/datadog: Update Datadog client dependency to v1.14.0 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * plugin/apm/prometheus: Update Prometheus client dependency to v1.12.2 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * plugin/target/aws: Update AWS client dependency to v1.16.4 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * plugin/target/azure: Update Azure client dependency to v64.1.0 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+ * plugin/target/gcp: Update GCP client dependency to 0.80.0 [[GH-585](https://github.com/hashicorp/nomad-autoscaler/pull/585)]
+
 ## 0.3.6 (February 18, 2022)
 
 IMPROVEMENTS: